+++ swov1/src/models/components/demprobe.ts
@@ -38,6 +38,8 @@ export type Coordinates = {
 export type DemProbe = {
   /**
    * Unique identifier of the probe.
+   *
+   * @deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
    */
   id: string;
   /**